requirements.txt
yfinance>=0.2.28
praw>=7.7.1
newsapi-python>=0.2.7
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
spacy>=3.7.0
neo4j>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.28.0
setuptools>=68.0.0
wheel>=0.42.0
huggingface_hub>=0.16.0
fastapi>=0.115.0
uvicorn>=0.30.0
google-genai>=1.0.0
redis>=5.0.0

